Remove bold username formatting from changes list
Seems like Slack hasn't decided yet if they want to render styles in context blocks or not. Bold text used to work in context blocks, and now it doesn't.

Let's keep formatting in context blocks to the minimum, and add it later if users ask for it + it actually works